Add files via upload
It is reported that rich countries are hoarding more Covid vaccines than needed for boosters while less wealthy countries are relying on a vaccine-sharing arrangement called Covax. This has led to a striking divide between regions of the world. Therefore it is insightful to analyse and observe that whether "Fully Vaccinated Percentage of the Country"( Vaccination Ranking list was taken from NewYorkTimes - Reference Link ) has affected to the number of cases, deaths and also whether there is any relation to covid symptoms.
Out of the ranked list, top 5 vaccinated countries and least 5 fully vaccinated countries are taken to analyse the confirmed cases and confirmed deaths in them.